# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "InvoiceTool Swagger API",
        "description": "InvoiceTool Swagger API description",
        "contact": {
            "email": "admin@finpay.com"
        },
        "version": "1.0.0"
    },
    "host": "invoice-tool-nov8.test/",
    "basePath": "api/v1",
    "paths": {
...
        "/Client": {
            "get": {
                "tags": [
                    "CLIENTS"
                ],
                "summary": "Clients Table",
                "operationId": "cid",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/client": {
            "post": {
                "tags": [
                    "CLIENTS"
                ],
                "summary": "Client Store",
                "description": "Client Store",
                "parameters": [
                    {
                        "name": "client_name",
                        "in": "formData",
                        "description": "Client name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "City",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "formData",
                        "description": "State",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "postal_code",
                        "in": "formData",
                        "description": "Postal Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "description": "Country",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone_number",
                        "in": "formData",
                        "description": "Phone Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fax_number",
                        "in": "formData",
                        "description": "Fax Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "web_site",
                        "in": "formData",
                        "description": "Website",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "default_currency",
                        "in": "formData",
                        "description": "Default Currency",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tax_code",
                        "in": "formData",
                        "description": "Tax Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "description": "Address",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Status",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/Client/{id}": {
            "get": {
                "tags": [
                    "CLIENTS"
                ],
                "summary": "Clients edit",
                "operationId": "Ceid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CLIENTS"
                ],
                "summary": "Clients Update",
                "description": "Clients Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Clients Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "client_name",
                        "in": "formData",
                        "description": "Client name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "City",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "state",
                        "in": "formData",
                        "description": "State",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "postal_code",
                        "in": "formData",
                        "description": "Postal Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "description": "Country",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone_number",
                        "in": "formData",
                        "description": "Phone Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fax_number",
                        "in": "formData",
                        "description": "Fax Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mobile_number",
                        "in": "formData",
                        "description": "Mobile Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "web_site",
                        "in": "formData",
                        "description": "Website",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "default_currency",
                        "in": "formData",
                        "description": "Default Currency",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tax_code",
                        "in": "formData",
                        "description": "Tax Number",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "description": "Address",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Status",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CLIENTS"
                ],
                "summary": "Client Delete",
                "operationId": "Cid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client Delete",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}